+++ Dnn.CommunityForums/Deprecated/LikeInfo.cs
@@ -16,22 +16,11 @@
 // THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF 
 // CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
 // DEALINGS IN THE SOFTWARE.
-//
- 
-using DotNetNuke.ComponentModel.DataAnnotations; 
+// 
+using System;
 using System.Web.Caching;
-
-namespace DotNetNuke.Modules.ActiveForums.Entities
-{
-    [TableName("activeforums_Likes")]
-    [PrimaryKey("Id", AutoIncrement = true)]
-    [Scope("PostId")]
-    [Cacheable("activeforums_Likes", CacheItemPriority.Normal)]
-    internal partial class LikeInfo
-    {
-        public int Id { get; set; }
-        public int PostId { get; set; }
-        public int UserId { get; set; }
-        public bool Checked { get; set; }
-    }
-}
+namespace DotNetNuke.Modules.ActiveForums
+{ 
+    [Obsolete("Deprecated in Community Forums. Scheduled removal in 09.00.00. Replace with DotNetNuke.Modules.ActiveForums.Entities.Likes")]
+    class Likes : DotNetNuke.Modules.ActiveForums.Entities.LikeInfo { }
+}
