+++ Dnn.CommunityForums/DnnCommunityForums.csproj
@@ -1123,9 +1123,10 @@
     <Compile Include="Controllers\TopicController.cs" />
     <Compile Include="Controllers\UrlController.cs" />
     <Compile Include="Controllers\UserController.cs" />
+    <Compile Include="Deprecated\LikeController.cs" />
+    <Compile Include="Deprecated\LikeInfo.cs" />
     <Compile Include="Entities\LikeInfo.cs" />
     <Compile Include="Entities\FilterInfo.cs" />
-    <Compile Include="Entities\Like.cs" />
     <Compile Include="Controllers\LikeController.cs" />
     <Compile Include="class\ParamBuilder.cs" />
     <Compile Include="class\StringExtensions.cs" />
